+++ CHANGELOG.md
@@ -115,6 +115,7 @@
 - Enh #116: Add `HasRefX` trait and corresponding tests for managing SVG `refX` attribute (@terabytesoftw)
 - Enh #117: Add `HasRefY` trait and corresponding tests for managing SVG `refY` attribute (@terabytesoftw)
 - Enh #118: Add `HasViewBox` trait and corresponding tests for managing SVG `viewBox` attribute (@terabytesoftw)
+- Enh #119: Add `Marker` class and corresponding tests for SVG `<marker>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024